# contracts/runbook-orchestrator.openapi.yaml
openapi: 3.1.0
info:
  title: AgentArmy Runbook Orchestrator — Control API
  version: 0.1.0
  description: >-
    Serve-mode control surface for the AgentArmy runbook-orchestrator
    (`templates/runbook-orchestrator-image/`). The orchestrator indexes a
    directory of runbook files (BPMN 2.0 XML + OASIS CACAO 2.0 JSON), subscribes
    each event-triggered runbook to a `fleet.*` NATS subject, and runs it on a
    matching CloudEvent. This API lists the index and lets an operator/agent
    fire a runbook on demand. The trigger surface itself (consumed `fleet.*`
    events; emitted `fleet.runbook.command` / `fleet.runbook.completed`) is the
    event-bus contract, governed by ARC-ADR-022 + ARC-ADR-031.
servers:
  - url: "{ORCHESTRATOR_URL}"
    description: Cluster-internal base; the orchestrator listens on PORT (default 8080).
    variables:
      ORCHESTRATOR_URL:
        default: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: healthz
      summary: Liveness — index size + subscribed trigger subjects
      responses:
        "200":
          description: Healthy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
  /runbooks:
    get:
      operationId: listRunbooks
      summary: List indexed runbooks
      responses:
        "200":
          description: The current runbook index.
          content:
            application/json:
              schema:
                type: object
                required: [runbooks]
                properties:
                  runbooks:
                    type: array
                    items: { $ref: "#/components/schemas/RunbookEntry" }
  /runbooks/{runbook_id}/trigger:
    post:
      operationId: triggerRunbook
      summary: Run a runbook now
      description: >-
        Executes the runbook to completion using the safe-orchestrator executor.
        The JSON request body becomes the execution context (variables read by
        conditions / switch steps).
      parameters:
        - in: path
          name: runbook_id
          required: true
          schema: { type: string }
          description: The runbook id (CACAO playbook id or BPMN process id).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Execution context (variable name → value).
              example: { severity: "HIGH" }
      responses:
        "200":
          description: Runbook executed (see status).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunSummary" }
        "404":
          description: Unknown runbook id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Runbook is structurally invalid (will not run).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  schemas:
    Health:
      type: object
      required: [status, runbooks, subscribed_subjects]
      properties:
        status: { type: string, const: ok }
        runbooks: { type: integer, description: Number of indexed runbooks. }
        subscribed_subjects:
          type: array
          items: { type: string }
          example: ["fleet.siem.phishing", "fleet.runbook.securityalertreceived"]
        nats:
          type: [string, "null"]
          description: NATS URL if connected; null when degraded to control-API-only.
    RunbookEntry:
      type: object
      required: [id, name, format, trigger, valid]
      properties:
        id: { type: string }
        name: { type: string }
        format: { type: string, enum: [bpmn, cacao] }
        trigger:
          type: object
          additionalProperties: true
          description: Start-trigger metadata (kind = message|timer|signal|none).
        valid: { type: boolean }
        errors:
          type: array
          items: { type: string }
        path: { type: string }
    RunSummary:
      type: object
      required: [runbook, format, status, steps]
      properties:
        runbook: { type: string }
        name: { type: string }
        format: { type: string, enum: [bpmn, cacao] }
        status: { type: string, enum: [completed, error] }
        steps: { type: integer, description: Number of trace records emitted. }
        error: { type: [string, "null"] }
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
          example: "unknown runbook 'playbook--…'"